requirements.txt
Django==3.1.13
django-rest-auth==0.9.5
django-rest-email-auth==3.0.3
djangorestframework==3.12.4
django-environ==0.4.5
psycopg2-binary==2.9.1
pillow==8.3.0